Improve Makefile

+++ Makefile
@@ -1,6 +1,8 @@
 # alternatively use ln -s to symlink the files instead of copying them
 INST = cp -rf
 
+clonevimplugin = test -e ~/.vim/bundle/$(1) || git clone $(2) ~/.vim/bundle/$(1)
+
 all: vim bc bashrc w3m nano browserweiche lesspipe xmodmap mpv
 
 nano:
@@ -27,22 +29,26 @@ vimplugins:
 	mkdir -p ~/.vim/autoload
 	mkdir -p ~/.vim/bundle
 	curl -LSso ~/.vim/autoload/pathogen.vim https://tpo.pe/pathogen.vim	
-	test -e ~/.vim/bundle/lightline || git clone https://github.com/itchyny/lightline.vim ~/.vim/bundle/lightline
-	test -e ~/.vim/bundle/nerdtree || git clone git://github.com/scrooloose/nerdtree.git ~/.vim/bundle/nerdtree
-	test -e ~/.vim/bundle/tabular || git clone git://github.com/godlygeek/tabular.git ~/.vim/bundle/tabular
-	test -e ~/.vim/bundle/flake8 || git clone https://github.com/nvie/vim-flake8.git ~/.vim/bundle/flake8
-	test -e ~/.vim/bundle/fugitive || git clone git://github.com/tpope/vim-fugitive.git ~/.vim/bundle/fugitive
-	test -e ~/.vim/bundle/gitgutter || git clone git://github.com/airblade/vim-gitgutter.git ~/.vim/bundle/gitgutter
-	test -e ~/.vim/bundle/mundo || git clone https://github.com/simnalamburt/vim-mundo.git ~/.vim/bundle/mundo
-	test -e ~/.vim/bundle/sleuth || git clone git://github.com/tpope/vim-sleuth.git ~/.vim/bundle/sleuth
-	test -e ~/.vim/bundle/xterm-color-table || git clone https://github.com/guns/xterm-color-table.vim.git ~/.vim/bundle/xterm-color-table
-	test -e ~/.vim/bundle/mru || git clone https://github.com/vim-scripts/mru.vim.git ~/.vim/bundle/mru
+	$(call clonevimplugin,lightline,https://github.com/itchyny/lightline.vim)
+	$(call clonevimplugin,nerdtree,git://github.com/scrooloose/nerdtree.git)
+	$(call clonevimplugin,tabular,git://github.com/godlygeek/tabular.git)
+	$(call clonevimplugin,flake8,https://github.com/nvie/vim-flake8.git)
+	$(call clonevimplugin,fugitive,git://github.com/tpope/vim-fugitive.git)
+	$(call clonevimplugin,gitgutter,git://github.com/airblade/vim-gitgutter.git)
+	$(call clonevimplugin,mundo,https://github.com/simnalamburt/vim-mundo.git)
+	$(call clonevimplugin,sleuth,git://github.com/tpope/vim-sleuth.git)
+	$(call clonevimplugin,rust,git://github.com/rust-lang/rust.vim.git)
+	$(call clonevimplugin,xterm-color-table,https://github.com/guns/xterm-color-table.vim.git)
+	$(call clonevimplugin,mru,https://github.com/vim-scripts/mru.vim.git)
+	$(call clonevimplugin,YouCompleteMe,https://github.com/Valloric/YouCompleteMe)
+	cd ~/.vim/bundle/YouCompleteMe; git submodule update --init --recursive
+	~/.vim/bundle/YouCompleteMe/install.py --racer-completer
 
 vim: themes vimplugins
-	mkdir ~/.vim
-	mkdir ~/.vim/.backup ~/.vim/.swp ~/.vim/.undo
-	$(INST) .vimrc ~/.vim/vimrc
-	ln -s ~/.vim/vimrc ~/.vim/init.vim
+	mkdir -p ~/.vim
+	mkdir -p ~/.vim/.backup ~/.vim/.swp ~/.vim/.undo
+	$(INST) vimrc ~/.vim/vimrc
+	test -e ~/.vim/init.vim || ln -s ~/.vim/vimrc ~/.vim/init.vim
 
 browserweiche:
 	$(INST) browserweiche.sh ~/.browserweiche.sh
